Database.Beam.Haskell.Syntax

Description

Instances that allow us to use Haskell as a backend syntax. This allows us to use migrations defined a la SQL to generate a beam schema.

Mainly of interest to backends.

Unfortunately, we define some orphan Hashable instances that aren't defined for us in haskell-src-exts.

Documentation

newtype HsDbField

Constructors

HsDbField

Fields

buildHsDbField :: Type () -> Type ()

data HsConstraintDefinition

Constructors

HsConstraintDefinition

Fields

hsConstraintDefinitionConstraint :: HsExpr

newtype HsEntityName

Constructors

HsEntityName

Fields

getHsEntityName :: String

data HsImport

Constructors

HsImportAll

HsImportSome (Set (ImportSpec ()))

importSome :: Text -> [ImportSpec ()] -> HsImports

importTyNamed :: Text -> ImportSpec ()

importVarNamed :: Text -> ImportSpec ()

newtype HsImports

Constructors

HsImports (Map (ModuleName ()) HsImport)

data HsDataType

Constructors

HsDataType

Fields

hsDataTypeMigration :: HsExpr

hsDataTypeType :: HsType

hsDataTypeSerialized :: BeamSerializedDataType

data HsType

Constructors

HsType

Fields

hsTypeSyntax :: Type ()

hsTypeImports :: HsImports

data HsExpr

Constructors

HsExpr

Fields

hsExprSyntax :: Exp ()

hsExprImports :: HsImports

hsExprConstraints :: [Asst ()]

hsExprTypeVariables :: Set (Name ())

data HsColumnSchema

Constructors

HsColumnSchema

Fields

mkHsColumnSchema :: Text -> HsExpr

hsColumnSchemaType :: HsType

data HsDecl

Constructors

HsDecl

Fields

hsDeclSyntax :: Decl ()

hsDeclImports :: HsImports

hsDeclExports :: [ExportSpec ()]

data HsAction

Constructors

HsAction

Fields

hsSyntaxMigration :: [(Maybe (Pat ()), HsExpr)]

hsSyntaxEntities :: [HsEntity]

newtype HsBackendConstraint

Constructors

HsBackendConstraint

Fields

buildHsBackendConstraint :: Type () -> Asst ()

data HsBeamBackend f

Constructors

HsBeamBackendSingle HsType f

HsBeamBackendConstrained [HsBackendConstraint]

HsBeamBackendNone

data HsEntity

Constructors

HsEntity

Fields

hsEntityBackend :: HsBeamBackend HsExpr

hsEntityName :: HsEntityName

hsEntityDecls :: [HsDecl]

hsEntityDbDecl :: HsDbField

hsEntityExp :: HsExpr

newtype HsFieldLookup

Constructors

HsFieldLookup

Fields

hsFieldLookup :: Text -> Maybe (Text, Type ())

newtype HsTableConstraint

Constructors

HsTableConstraint (Text -> HsFieldLookup -> HsTableConstraintDecls)

data HsTableConstraintDecls

Constructors

HsTableConstraintDecls

Fields

hsTableConstraintInstance :: [InstDecl ()]

hsTableConstraintDecls :: [HsDecl]

data HsModule

Constructors

HsModule

Fields

hsModuleName :: String

hsModuleEntities :: [HsEntity]

hsModuleMigration :: [(Maybe (Pat ()), HsExpr)]

hsActionsToModule :: String -> [HsAction] -> HsModule

unqual :: String -> QName ()

entityDbFieldName :: HsEntity -> String

derivingDecl :: [InstRule ()] -> Deriving ()

dataDecl :: DeclHead () -> [QualConDecl ()] -> Maybe (Deriving ()) -> Decl ()

insDataDecl :: Type () -> [QualConDecl ()] -> Maybe (Deriving ()) -> InstDecl ()

databaseTypeDecl :: [HsEntity] -> Decl ()

migrationTypeDecl :: HsBeamBackend HsExpr -> [Type ()] -> Decl ()

migrationDecl :: HsBeamBackend HsExpr -> [Exp ()] -> [(Maybe (Pat ()), HsExpr)] -> [HsEntity] -> Decl ()

dbTypeDecl :: HsBeamBackend HsExpr -> Decl ()

dbDecl :: HsBeamBackend HsExpr -> [HsExpr] -> Decl ()

renderHsSchema :: HsModule -> Either String String

DDL Syntax definitions

data HsNone

Constructors

HsNone

data HsMigrateBackend

Constructors

HsMigrateBackend

hsMkTableName :: (Char -> Char) -> TableName -> String

hsTableVarName :: TableName -> String

hsTableTypeName :: TableName -> String

hsErrorType :: String -> HsDataType

HsSyntax utilities

tyParens :: Type () -> Type ()

functionTy :: Type () -> Type () -> Type ()

tyTuple :: [Type ()] -> Type ()

tyApp :: Type () -> [Type ()] -> Type ()

tyConNamed :: String -> Type ()

tyVarNamed :: String -> Type ()

combineHsExpr :: (Exp () -> Exp () -> Exp ()) -> HsExpr -> HsExpr -> HsExpr

hsApp :: HsExpr -> [HsExpr] -> HsExpr

hsVisibleTyApp :: HsExpr -> Type () -> HsExpr

hsApApp :: HsExpr -> [HsExpr] -> HsExpr

hsStr :: Text -> HsExpr

hsRecCon :: Text -> [(Text, HsExpr)] -> HsExpr

hsMaybe :: Maybe HsExpr -> HsExpr

hsVar :: Text -> HsExpr

hsVarFrom :: Text -> Text -> HsExpr

hsTyCon :: Text -> HsExpr

hsInt :: (Integral a, Show a) => a -> HsExpr

hsOp :: Text -> QOp ()

hsInstance :: Text -> [Type ()] -> [InstDecl ()] -> Decl ()

hsDerivingInstance :: Text -> [Type ()] -> Decl ()

hsList :: [HsExpr] -> HsExpr

hsTuple :: [HsExpr] -> HsExpr

inst :: String -> InstRule ()

beamMigrateSqlBackend :: HsBackendConstraint

Orphans

Orphan instances

Hashable Boxed

Instance details

Methods

hashWithSalt :: Int -> Boxed -> Int

hash :: Boxed -> Int

Hashable (ModuleName ())

Instance details

Methods

hashWithSalt :: Int -> ModuleName () -> Int

hash :: ModuleName () -> Int

Hashable (SpecialCon ())

Instance details

Methods

hashWithSalt :: Int -> SpecialCon () -> Int

hash :: SpecialCon () -> Int

Hashable (QName ())

Instance details

Methods

hashWithSalt :: Int -> QName () -> Int

hash :: QName () -> Int

Hashable (Name ())

Instance details

Methods

hashWithSalt :: Int -> Name () -> Int

hash :: Name () -> Int

Hashable (IPName ())

Instance details

Methods

hashWithSalt :: Int -> IPName () -> Int

hash :: IPName () -> Int

Hashable (QOp ())

Instance details

Methods

hashWithSalt :: Int -> QOp () -> Int

hash :: QOp () -> Int

Hashable (Op ())

Instance details

Methods

hashWithSalt :: Int -> Op () -> Int

hash :: Op () -> Int

Hashable (CName ())

Instance details

Methods

hashWithSalt :: Int -> CName () -> Int

hash :: CName () -> Int

Hashable (Namespace ())

Instance details

Methods

hashWithSalt :: Int -> Namespace () -> Int

hash :: Namespace () -> Int

Hashable (ImportSpec ())

Instance details

Methods

hashWithSalt :: Int -> ImportSpec () -> Int

hash :: ImportSpec () -> Int

Hashable (Assoc ())

Instance details

Methods

hashWithSalt :: Int -> Assoc () -> Int

hash :: Assoc () -> Int

Hashable (Decl ())

Instance details

Methods

hashWithSalt :: Int -> Decl () -> Int

hash :: Decl () -> Int

Hashable (PatternSynDirection ())

Instance details

Methods

hashWithSalt :: Int -> PatternSynDirection () -> Int

hash :: PatternSynDirection () -> Int

Hashable (TypeEqn ())

Instance details

Methods

hashWithSalt :: Int -> TypeEqn () -> Int

hash :: TypeEqn () -> Int

Hashable (Annotation ())

Instance details

Methods

hashWithSalt :: Int -> Annotation () -> Int

hash :: Annotation () -> Int

Hashable (BooleanFormula ())

Instance details

Methods

hashWithSalt :: Int -> BooleanFormula () -> Int

hash :: BooleanFormula () -> Int

Hashable (Role ())

Instance details

Methods

hashWithSalt :: Int -> Role () -> Int

hash :: Role () -> Int

Hashable (DataOrNew ())

Instance details

Methods

hashWithSalt :: Int -> DataOrNew () -> Int

hash :: DataOrNew () -> Int

Hashable (InjectivityInfo ())

Instance details

Methods

hashWithSalt :: Int -> InjectivityInfo () -> Int

hash :: InjectivityInfo () -> Int

Hashable (ResultSig ())

Instance details

Methods

hashWithSalt :: Int -> ResultSig () -> Int

hash :: ResultSig () -> Int

Hashable (DeclHead ())

Instance details

Methods

hashWithSalt :: Int -> DeclHead () -> Int

hash :: DeclHead () -> Int

Hashable (InstRule ())

Instance details

Methods

hashWithSalt :: Int -> InstRule () -> Int

hash :: InstRule () -> Int

Hashable (InstHead ())

Instance details

Methods

hashWithSalt :: Int -> InstHead () -> Int

hash :: InstHead () -> Int

Hashable (Deriving ())

Instance details

Methods

hashWithSalt :: Int -> Deriving () -> Int

hash :: Deriving () -> Int

Hashable (DerivStrategy ())

Instance details

Methods

hashWithSalt :: Int -> DerivStrategy () -> Int

hash :: DerivStrategy () -> Int

Hashable (Binds ())

Instance details

Methods

hashWithSalt :: Int -> Binds () -> Int

hash :: Binds () -> Int

Hashable (IPBind ())

Instance details

Methods

hashWithSalt :: Int -> IPBind () -> Int

hash :: IPBind () -> Int

Hashable (Match ())

Instance details

Methods

hashWithSalt :: Int -> Match () -> Int

hash :: Match () -> Int

Hashable (QualConDecl ())

Instance details

Methods

hashWithSalt :: Int -> QualConDecl () -> Int

hash :: QualConDecl () -> Int

Hashable (ConDecl ())

Instance details

Methods

hashWithSalt :: Int -> ConDecl () -> Int

hash :: ConDecl () -> Int

Hashable (FieldDecl ())

Instance details

Methods

hashWithSalt :: Int -> FieldDecl () -> Int

hash :: FieldDecl () -> Int

Hashable (GadtDecl ())

Instance details

Methods

hashWithSalt :: Int -> GadtDecl () -> Int

hash :: GadtDecl () -> Int

Hashable (ClassDecl ())

Instance details

Methods

hashWithSalt :: Int -> ClassDecl () -> Int

hash :: ClassDecl () -> Int

Hashable (InstDecl ())

Instance details

Methods

hashWithSalt :: Int -> InstDecl () -> Int

hash :: InstDecl () -> Int

Hashable (BangType ())

Instance details

Methods

hashWithSalt :: Int -> BangType () -> Int

hash :: BangType () -> Int

Hashable (Unpackedness ())

Instance details

Methods

hashWithSalt :: Int -> Unpackedness () -> Int

hash :: Unpackedness () -> Int

Hashable (Rhs ())

Instance details

Methods

hashWithSalt :: Int -> Rhs () -> Int

hash :: Rhs () -> Int

Hashable (GuardedRhs ())

Instance details

Methods

hashWithSalt :: Int -> GuardedRhs () -> Int

hash :: GuardedRhs () -> Int

Hashable (Type ())

Instance details

Methods

hashWithSalt :: Int -> Type () -> Int

hash :: Type () -> Int

Hashable (MaybePromotedName ())

Instance details

Methods

hashWithSalt :: Int -> MaybePromotedName () -> Int

hash :: MaybePromotedName () -> Int

Hashable (Promoted ())

Instance details

Methods

hashWithSalt :: Int -> Promoted () -> Int

hash :: Promoted () -> Int

Hashable (TyVarBind ())

Instance details

Methods

hashWithSalt :: Int -> TyVarBind () -> Int

hash :: TyVarBind () -> Int

Hashable (FunDep ())

Instance details

Methods

hashWithSalt :: Int -> FunDep () -> Int

hash :: FunDep () -> Int

Hashable (Context ())

Instance details

Methods

hashWithSalt :: Int -> Context () -> Int

hash :: Context () -> Int

Hashable (Asst ())

Instance details

Methods

hashWithSalt :: Int -> Asst () -> Int

hash :: Asst () -> Int

Hashable (Literal ())

Instance details

Methods

hashWithSalt :: Int -> Literal () -> Int

hash :: Literal () -> Int

Hashable (Sign ())

Instance details

Methods

hashWithSalt :: Int -> Sign () -> Int

hash :: Sign () -> Int

Hashable (Exp ())

Instance details

Methods

hashWithSalt :: Int -> Exp () -> Int

hash :: Exp () -> Int

Hashable (XName ())

Instance details

Methods

hashWithSalt :: Int -> XName () -> Int

hash :: XName () -> Int

Hashable (XAttr ())

Instance details

Methods

hashWithSalt :: Int -> XAttr () -> Int

hash :: XAttr () -> Int

Hashable (Bracket ())

Instance details

Methods

hashWithSalt :: Int -> Bracket () -> Int

hash :: Bracket () -> Int

Hashable (Splice ())

Instance details

Methods

hashWithSalt :: Int -> Splice () -> Int

hash :: Splice () -> Int

Hashable (Safety ())

Instance details

Methods

hashWithSalt :: Int -> Safety () -> Int

hash :: Safety () -> Int

Hashable (CallConv ())

Instance details

Methods

hashWithSalt :: Int -> CallConv () -> Int

hash :: CallConv () -> Int

Hashable (Overlap ())

Instance details

Methods

hashWithSalt :: Int -> Overlap () -> Int

hash :: Overlap () -> Int

Hashable (Activation ())

Instance details

Methods

hashWithSalt :: Int -> Activation () -> Int

hash :: Activation () -> Int

Hashable (Rule ())

Instance details

Methods

hashWithSalt :: Int -> Rule () -> Int

hash :: Rule () -> Int

Hashable (RuleVar ())

Instance details

Methods

hashWithSalt :: Int -> RuleVar () -> Int

hash :: RuleVar () -> Int

Hashable (Pat ())

Instance details

Methods

hashWithSalt :: Int -> Pat () -> Int

hash :: Pat () -> Int

Hashable (PXAttr ())

Instance details

Methods

hashWithSalt :: Int -> PXAttr () -> Int

hash :: PXAttr () -> Int

Hashable (RPatOp ())

Instance details

Methods

hashWithSalt :: Int -> RPatOp () -> Int

hash :: RPatOp () -> Int

Hashable (RPat ())

Instance details

Methods

hashWithSalt :: Int -> RPat () -> Int

hash :: RPat () -> Int

Hashable (PatField ())

Instance details

Methods

hashWithSalt :: Int -> PatField () -> Int

hash :: PatField () -> Int

Hashable (Stmt ())

Instance details

Methods

hashWithSalt :: Int -> Stmt () -> Int

hash :: Stmt () -> Int

Hashable (QualStmt ())

Instance details

Methods

hashWithSalt :: Int -> QualStmt () -> Int

hash :: QualStmt () -> Int

Hashable (FieldUpdate ())

Instance details

Methods

hashWithSalt :: Int -> FieldUpdate () -> Int

hash :: FieldUpdate () -> Int

Hashable (Alt ())

Instance details

Methods

hashWithSalt :: Int -> Alt () -> Int

hash :: Alt () -> Int