Database.Beam.Migrate.Checks

Description

Defines common DatabasePredicates that are shared among backends

Table checks

data TableExistsPredicate

Asserts that a table with the given name exists in a database

Constructors

TableExistsPredicate QualifiedName
Table name

class HasDataTypeCreatedCheck dataType where

A class that can check whether a particular data type is present in a set of preconditions.

Methods

dataTypeHasBeenCreated :: dataType -> (forall preCondition. Typeable preCondition => [preCondition]) -> Bool

data TableHasColumn be where

Asserts that the table specified has a column with the given data type. The type paramater syntax should be an instance of IsSql92ColumnSchemaSyntax.

Constructors

TableHasColumn

Fields

:: HasDataTypeCreatedCheck (BeamMigrateSqlBackendDataTypeSyntax be)

=> { hasColumn_table :: QualifiedName

Table name

, hasColumn_column :: Text

Column name

, hasColumn_type :: BeamMigrateSqlBackendDataTypeSyntax be

Data type

} -> TableHasColumn be

data TableColumnHasConstraint be

Asserts that a particular column of a table has a given constraint. The syntax type parameter should be an instance of IsSql92ColumnSchemaSyntax

Constructors

TableColumnHasConstraint

Fields

hasConstraint_table :: QualifiedName

Table name

hasConstraint_column :: Text

Column name

hasConstraint_defn :: BeamSqlBackendColumnConstraintDefinitionSyntax be

Constraint definition

data TableHasPrimaryKey

Asserts that the given table has a primary key made of the given columns. The order of the columns is significant.

Constructors

TableHasPrimaryKey

Fields

hasPrimaryKey_table :: QualifiedName

Table name

hasPrimaryKey_cols :: [Text]

Column names

Deserialization

beamCheckDeserializers :: forall be. (Typeable be, BeamMigrateOnlySqlBackend be, HasDataTypeCreatedCheck (BeamMigrateSqlBackendDataTypeSyntax be)) => BeamDeserializers be

BeamDeserializers for all the predicates defined in this module