Database.Beam.Migrate.Generics

Description

Support for creating checked databases from Haskell ADTs, using Generics.

For more information, see the manual

Default checked database settings

defaultMigratableDbSettings :: forall be db. (Generic (CheckedDatabaseSettings be db), GAutoMigratableDb be (Rep (CheckedDatabaseSettings be db))) => CheckedDatabaseSettings be db

Produce a checked database for the given Haskell database type

See the manual for more information on the defaults.

Extending the defaulting sytem

class BeamMigrateSqlBackend be => HasDefaultSqlDataType be ty where

Used to define a default SQL data type for a haskell type in a particular backend, as well as any constraints that are needed

Beam defines instances for several standard SQL types, which are polymorphic over any standard data type syntax. Backends or extensions which provide custom types should instantiate instances of this class for any types they provide for which they would like checked schema migrations

Minimal complete definition

defaultSqlDataType

Methods

defaultSqlDataType

Arguments

:: Proxy ty
Concrete representation of the type

-> Proxy be
Concrete representation of the backend

-> Bool
True if this field is in an embedded key or table, False otherwise

-> BeamSqlBackendDataTypeSyntax be

Provide a data type for the given type

defaultSqlDataTypeConstraints

Arguments

:: Proxy ty
Concrete representation of the type

-> Proxy be
Concrete representation of the backend

-> Bool
True if this field is embedded in a foreign key, False otherwise. For example, SERIAL types in postgres get a DEFAULT constraint, but SERIAL types in a foreign key do not.

-> [FieldCheck]

Provide arbitrary constraints on a field of the requested type. See FieldCheck for more information on the formatting of constraints.

class BeamMigrateSqlBackend be => HasNullableConstraint (x :: Bool) be

Minimal complete definition

nullableConstraint

type family NullableStatus (x :: *) :: Bool where ...

Equations

NullableStatus (Maybe x) = 'True

NullableStatus x = 'False