Database.Beam.Migrate.SQL.Builder

Description

DDL syntax instances for SqlSyntaxBuilder

Documentation

data SqlSyntaxBuilderCreateTableOptions

Options for CREATE TABLE. Given as a separate ADT because the options may go in different places syntactically.

You never really need to use this type directly.

Constructors

SqlSyntaxBuilderCreateTableOptions SqlSyntaxBuilder SqlSyntaxBuilder

data ConstraintAttributeTiming

Some backends use this to represent their constraint attributes. Does not need to be used in practice.

Constructors

InitiallyDeferred

InitiallyImmediate

data SqlConstraintAttributesBuilder

Valid IsSql92ConstraintAttributesSyntax shared among some backends.

Constructors

SqlConstraintAttributesBuilder

Fields

_sqlConstraintAttributeTiming :: Maybe ConstraintAttributeTiming

_sqlConstraintAttributeDeferrable :: Maybe Bool

fromSqlConstraintAttributes :: SqlConstraintAttributesBuilder -> Builder

Convert a SqlConstraintAttributesBuilder to its SQL92 representation in the returned ByteString Builder.

sqlConstraintAttributesSerialized :: SqlConstraintAttributesBuilder -> BeamSerializedConstraintAttributes

Serialize a SqlConstraintAttributesBuilder

Orphan instances

Sql92DisplaySyntax SqlSyntaxBuilder

Instance details

Methods

displaySyntax :: SqlSyntaxBuilder -> String

IsSql92ColumnConstraintSyntax SqlSyntaxBuilder

Instance details

Associated Types

type Sql92ColumnConstraintMatchTypeSyntax SqlSyntaxBuilder

type Sql92ColumnConstraintReferentialActionSyntax SqlSyntaxBuilder

type Sql92ColumnConstraintExpressionSyntax SqlSyntaxBuilder

Methods

notNullConstraintSyntax :: SqlSyntaxBuilder

uniqueColumnConstraintSyntax :: SqlSyntaxBuilder

primaryKeyColumnConstraintSyntax :: SqlSyntaxBuilder

checkColumnConstraintSyntax :: Sql92ColumnConstraintExpressionSyntax SqlSyntaxBuilder -> SqlSyntaxBuilder

referencesConstraintSyntax :: Text -> [Text] -> Maybe (Sql92ColumnConstraintMatchTypeSyntax SqlSyntaxBuilder) -> Maybe (Sql92ColumnConstraintReferentialActionSyntax SqlSyntaxBuilder) -> Maybe (Sql92ColumnConstraintReferentialActionSyntax SqlSyntaxBuilder) -> SqlSyntaxBuilder

IsSql92ColumnConstraintDefinitionSyntax SqlSyntaxBuilder

Instance details

Associated Types

type Sql92ColumnConstraintDefinitionConstraintSyntax SqlSyntaxBuilder

type Sql92ColumnConstraintDefinitionAttributesSyntax SqlSyntaxBuilder

Methods

constraintDefinitionSyntax :: Maybe Text -> Sql92ColumnConstraintDefinitionConstraintSyntax SqlSyntaxBuilder -> Maybe (Sql92ColumnConstraintDefinitionAttributesSyntax SqlSyntaxBuilder) -> SqlSyntaxBuilder

IsSql92ReferentialActionSyntax SqlSyntaxBuilder

Instance details

Methods

referentialActionCascadeSyntax :: SqlSyntaxBuilder

referentialActionSetNullSyntax :: SqlSyntaxBuilder

referentialActionSetDefaultSyntax :: SqlSyntaxBuilder

referentialActionNoActionSyntax :: SqlSyntaxBuilder

IsSql92MatchTypeSyntax SqlSyntaxBuilder

Instance details

Methods

fullMatchSyntax :: SqlSyntaxBuilder

partialMatchSyntax :: SqlSyntaxBuilder

IsSql92TableConstraintSyntax SqlSyntaxBuilder

Instance details

Methods

primaryKeyConstraintSyntax :: [Text] -> SqlSyntaxBuilder

IsSql92ColumnSchemaSyntax SqlSyntaxBuilder

Instance details

Associated Types

type Sql92ColumnSchemaColumnTypeSyntax SqlSyntaxBuilder

type Sql92ColumnSchemaExpressionSyntax SqlSyntaxBuilder

type Sql92ColumnSchemaColumnConstraintDefinitionSyntax SqlSyntaxBuilder

Methods

columnSchemaSyntax :: Sql92ColumnSchemaColumnTypeSyntax SqlSyntaxBuilder -> Maybe (Sql92ColumnSchemaExpressionSyntax SqlSyntaxBuilder) -> [Sql92ColumnSchemaColumnConstraintDefinitionSyntax SqlSyntaxBuilder] -> Maybe Text -> SqlSyntaxBuilder

IsSql92AlterColumnActionSyntax SqlSyntaxBuilder

Instance details

Methods

setNotNullSyntax :: SqlSyntaxBuilder

setNullSyntax :: SqlSyntaxBuilder

IsSql92AlterTableActionSyntax SqlSyntaxBuilder

Instance details

Associated Types

type Sql92AlterTableAlterColumnActionSyntax SqlSyntaxBuilder

type Sql92AlterTableColumnSchemaSyntax SqlSyntaxBuilder

Methods

alterColumnSyntax :: Text -> Sql92AlterTableAlterColumnActionSyntax SqlSyntaxBuilder -> SqlSyntaxBuilder

addColumnSyntax :: Text -> Sql92AlterTableColumnSchemaSyntax SqlSyntaxBuilder -> SqlSyntaxBuilder

dropColumnSyntax :: Text -> SqlSyntaxBuilder

renameTableToSyntax :: Text -> SqlSyntaxBuilder

renameColumnToSyntax :: Text -> Text -> SqlSyntaxBuilder

IsSql92AlterTableSyntax SqlSyntaxBuilder

Instance details

Associated Types

type Sql92AlterTableTableNameSyntax SqlSyntaxBuilder

type Sql92AlterTableAlterTableActionSyntax SqlSyntaxBuilder

Methods

alterTableSyntax :: Sql92AlterTableTableNameSyntax SqlSyntaxBuilder -> Sql92AlterTableAlterTableActionSyntax SqlSyntaxBuilder -> SqlSyntaxBuilder

IsSql92DropTableSyntax SqlSyntaxBuilder

Instance details

Associated Types

type Sql92DropTableTableNameSyntax SqlSyntaxBuilder

Methods

dropTableSyntax :: Sql92DropTableTableNameSyntax SqlSyntaxBuilder -> SqlSyntaxBuilder

IsSql92CreateTableSyntax SqlSyntaxBuilder

Instance details

Associated Types

type Sql92CreateTableTableNameSyntax SqlSyntaxBuilder

type Sql92CreateTableColumnSchemaSyntax SqlSyntaxBuilder

type Sql92CreateTableTableConstraintSyntax SqlSyntaxBuilder

type Sql92CreateTableOptionsSyntax SqlSyntaxBuilder

Methods

createTableSyntax :: Maybe (Sql92CreateTableOptionsSyntax SqlSyntaxBuilder) -> Sql92CreateTableTableNameSyntax SqlSyntaxBuilder -> [(Text, Sql92CreateTableColumnSchemaSyntax SqlSyntaxBuilder)] -> [Sql92CreateTableTableConstraintSyntax SqlSyntaxBuilder] -> SqlSyntaxBuilder

IsSql92DdlCommandSyntax SqlSyntaxBuilder

Instance details

Associated Types

type Sql92DdlCommandCreateTableSyntax SqlSyntaxBuilder

type Sql92DdlCommandAlterTableSyntax SqlSyntaxBuilder

type Sql92DdlCommandDropTableSyntax SqlSyntaxBuilder

Methods

createTableCmd :: Sql92DdlCommandCreateTableSyntax SqlSyntaxBuilder -> SqlSyntaxBuilder

dropTableCmd :: Sql92DdlCommandDropTableSyntax SqlSyntaxBuilder -> SqlSyntaxBuilder

alterTableCmd :: Sql92DdlCommandAlterTableSyntax SqlSyntaxBuilder -> SqlSyntaxBuilder